Move to subfolder and change some icons
