+++ src/subcomponents/footer/OpenTradeScreenButton.js
